refactor: move goreleaser/cli to one higher level
No need to keep cli as separate package
