+++ catalog/goreleaser/doc.go
@@ -1,0 +1,2 @@
+// Package goreleaser provides a set of utilities to create and run containers.
+package goreleaser
