+++ catalog/goreleaser/release/mage.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/magefile/mage/mg"
 
-	"github.com/mesosphere/daggers/catalog/goreleaser/cli"
+	"github.com/mesosphere/daggers/catalog/goreleaser"
 	"github.com/mesosphere/daggers/daggers"
 )
 
@@ -51,7 +51,7 @@ func ReleaseSnapshot(_ context.Context) error {
 // ReleaseWithOptions runs goreleaser release with specific options.
 //
 //nolint:revive // Disable stuttering check.
-func ReleaseWithOptions(opts ...daggers.Option[config]) (*cli.Result, error) {
+func ReleaseWithOptions(opts ...daggers.Option[config]) (*goreleaser.Result, error) {
 	debug := mg.Debug() || mg.Verbose()
 
 	options, err := daggers.InitConfig(opts...)
@@ -63,5 +63,5 @@ func ReleaseWithOptions(opts ...daggers.Option[config]) (*cli.Result, error) {
 		options = opt(options)
 	}
 
-	return cli.Run(cli.CommandRelease, debug, options.Env, options.toArgs())
+	return goreleaser.Run(goreleaser.CommandRelease, debug, options.Env, options.toArgs())
 }
